Refactor some classes for better code quality
Previously, the classes had long methods that were harder to read and do
not follow the SLAP and KISS principles closely. Furthermore, there had
been some unsafe practices such as the presence of dead code within the
project.

This affects the readability of the code and may introduce bugs. Hence,
through refactoring and removing unsafe portions of code, hopefully
these issues are minimized.